Add float instead of int finance_calculator.py

+++ python_introduction/finance_calculator.py
@@ -1,5 +1,5 @@
-income = int(input("Enter your monthly income: "))
-expense = int(input("Enter your total monthly expenses: "))
+income = float(input("Enter your monthly income: "))
+expense = float(input("Enter your total monthly expenses: "))
 interest = 0.05
 
 monthly_savings = income - expense
